Split everything to separate files
