+++ sourced/ml/transformers/__init__.py
@@ -1,0 +1,6 @@
+from sourced.ml.transformers.basic import *
+from sourced.ml.transformers.batch_transformers import \
+    BagsBatcher, BagsBatchSaver, BagsBatchParquetLoader, BagsBatch
+from sourced.ml.transformers.transformer import Transformer
+from sourced.ml.transformers.repo2df import Repo2DocFreq
+from sourced.ml.transformers.repo2weighted import Repo2WeightedSet
